requirements.txt
Flask==2.1.0
gunicorn==20.1.0
openai
langchain==0.0.232
pydantic
pandas
scikit-learn==1.3.0
pinecone-client
pinecone-client[grpc]==2.2.1
flask-cors
matplotlib